Move Notify code to a separate file
