+++ Samples/BackgroundBlur/BackgroundBlur/BackgroundBlur.vcxproj.filters
@@ -5,24 +5,25 @@
     <ClCompile Include="Capture.cpp" />
     <ClCompile Include="PreviewWnd.cpp" />
     <ClCompile Include="winmain.cpp" />
-    <ClCompile Include="Helpers\CSampleQueue.cpp" />
+    <ClCompile Include="External\CSampleQueue.cpp" />
     <ClCompile Include="TransformAsync.cpp" />
     <ClCompile Include="TransformAsync_IMFAsyncCallback.cpp" />
     <ClCompile Include="TransformAsync_IMFMediaEventGenerator.cpp" />
     <ClCompile Include="TransformAsync_IMFShutdown.cpp" />
     <ClCompile Include="TransformAsync_IMFTransform.cpp" />
-    <ClCompile Include="Helpers\utils.cpp" />
+    <ClCompile Include="External\utils.cpp" />
+    <ClCompile Include="TransformAsync_IMFVideoSampleAllocatorNotify.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="SegmentModel.h" />
     <ClInclude Include="Capture.h" />
     <ClInclude Include="Resource.h" />
-    <ClInclude Include="Helpers\CSampleQueue.h" />
+    <ClInclude Include="External\CSampleQueue.h" />
     <ClInclude Include="TransformAsync.h" />
-    <ClInclude Include="Helpers\common.h" />
-    <ClInclude Include="Helpers\logging.h" />
-    <ClInclude Include="Helpers\logmediatype.h" />
-    <ClInclude Include="Helpers\trace.h" />
+    <ClInclude Include="External\common.h" />
+    <ClInclude Include="External\logging.h" />
+    <ClInclude Include="External\logmediatype.h" />
+    <ClInclude Include="External\trace.h" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
