+++ Samples/BackgroundBlur/BackgroundBlur/Capture.h
@@ -32,7 +32,7 @@
 #include <commctrl.h>
 #include <d3d11.h>
 #include <initguid.h>
-#include "Helpers/common.h"
+#include "External/common.h"
 
 const UINT WM_APP_CAPTURE_EVENT = WM_APP + 1;
