+++ Samples/BackgroundBlur/BackgroundBlur/External/CSampleQueue.cpp
