+++ Samples/BackgroundBlur/BackgroundBlur/External/CSampleQueue.h
