+++ Samples/BackgroundBlur/BackgroundBlur/External/logging.h
