+++ Samples/BackgroundBlur/BackgroundBlur/External/logmediatype.h
