+++ Samples/BackgroundBlur/BackgroundBlur/External/trace.h
