+++ Samples/BackgroundBlur/BackgroundBlur/TransformAsync.cpp
@@ -1,20 +1,19 @@
 #include "TransformAsync.h"
-#include <winrt/windows.media.h>
-#include <winrt/windows.foundation.h>
 #include <windows.graphics.directx.direct3d11.interop.h>
 #include <winrt/Windows.Graphics.DirectX.Direct3D11.h>
 #include <windows.graphics.imaging.interop.h>
 #include <windows.media.core.interop.h>
-#include "Helpers/common.h"
+#include "External/common.h"
 #include <unknwn.h>
 #include <stdio.h>
 #include <wchar.h>
 #include <future>
 
 #define MFT_NUM_DEFAULT_ATTRIBUTES  4
-using namespace MainWindow;
+#define MIN_ALLOCATED_SAMPLES 2
+#define MAX_ALLOCATED_SAMPLES std::thread::hardware_concurrency()
 
-long long g_now; // The time since the last call to FrameThreadProc
+using namespace MainWindow;
 
 // Video FOURCC codes.
 const FOURCC FOURCC_NV12 = MAKEFOURCC('N', 'V', '1', '2');
@@ -281,46 +280,6 @@ HRESULT TransformAsync::FinishEval(com_ptr<IMFSample> pInputSample, com_ptr<IMFS
 }
 
 
-DWORD __stdcall FrameThreadProc(LPVOID lpParam) 
-{
-    DWORD waitResult; 
-    // Get the handle from the lpParam pointer
-    HANDLE event = lpParam;   
-    //OutputDebugString(L"Thread %d waiting for Frame event...");
-    waitResult = WaitForSingleObject(
-        event,         // event handle
-        INFINITE);      // indefinite wait
-   
-
-    switch (waitResult) {
-    case WAIT_OBJECT_0:
-        // TODO: Capture time and write to preview
-        if (g_now == NULL){
-            g_now = std::chrono::duration_cast<std::chrono::milliseconds>(std::chrono::high_resolution_clock::now().time_since_epoch()).count();
-            //OutputDebugString(L"First time responding to event!");
-        }
-        else {
-            auto l_now = std::chrono::duration_cast<std::chrono::milliseconds>(std::chrono::high_resolution_clock::now().time_since_epoch()).count();
-            auto timePassed = l_now - g_now;
-            g_now = l_now;
-            auto fps = 30000 / timePassed; // TODO: marco on num frames to update after? 
-            OutputDebugString(L"THREAD: ");
-            OutputDebugString(std::to_wstring(fps).c_str());
-            OutputDebugString(L"\n");
-
-            auto message = std::wstring(L"Frame Rate: ") + std::to_wstring(fps) + L" FPS";
-            MainWindow::_SetStatusText(message.c_str());
-            //TRACE(("Responded to event and it's been %d miliseconds", timePassed));
-            // TODO: Call Set status text with new framerate
-        }
-
-        break;
-    default: 
-        TRACE(("Wait error (%d)\n", GetLastError()));
-        return 0;
-    }
-    return 1;
-}
 
 HRESULT TransformAsync::OnSetD3DManager(ULONG_PTR ulParam)
 {
